added table reservation (with db caching)

+++ app/src/main/java/com/application/restaurantreservation/model/Customer.java
@@ -13,7 +13,7 @@ public class Customer {
     private String customerLastName;
     @SerializedName("id")
     @Expose
-    private long id;
+    private int id;
 
     public String getCustomerFirstName() {
         return customerFirstName;
@@ -31,11 +31,11 @@ public class Customer {
         this.customerLastName = customerLastName;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(int id) {
         this.id = id;
     }
